requirements.txt
allure-behave==2.13.2
allure-python-commons==2.13.2
Appium-Python-Client==2.8.1
selenium==4.9.0
